Update addFavorite to remove unused code
