+++ src/components/CreatureDetail.js
@@ -42,10 +42,10 @@ class CreatureDetail extends Component {
 
   render() {
     const creatureInfo = this.state.creatureDetail;
-    const actionsAll = this.state.creatureDetail.actions
-    console.log('creature detail:', this.state.creatureDetail)
-    console.log('type', this.state.creatureType);
-    console.log('actions', actionsAll)
+    // const actionsAll = this.state.creatureDetail.actions
+    // console.log('creature detail:', this.state.creatureDetail)
+    // console.log('type', this.state.creatureType);
+    // console.log('actions', actionsAll)
     // const actionsEach = this.state.creatureDetail.actions.forEach(action => { return action.name })
     // const actionsEach = this.state.creatureDetail.actions[0]
     return (
